Feature/ handled API part for private chat polls
